# web/docs/swagger.yaml
basePath: /
definitions:
  errorinterface.ErrorResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
...
  models.AuditLog:
    properties:
      action:
        description: Action taken (e.g., "Created Task", "Updated Milestone")
        type: string
      details:
        description: Detailed description of the action
        type: string
      timestamp:
        description: When the action was performed
        type: string
      user_id:
        description: The user who performed the action
        type: string
    type: object
...
  models.Workspace:
    properties:
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      members:
        items:
          type: string
        type: array
      name:
        description: Workspace details
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_timestamp_utc:
        type: string
    type: object
...
  workspaces.AddMemberRequest:
    properties:
      memberID:
        type: string
    type: object
  workspaces.WorkspaceRequest:
    description: Workspace request model
    properties:
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      members:
        items:
          type: string
        type: array
      name:
        description: Workspace details
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_timestamp_utc:
        type: string
    type: object
  workspaces.WorkspaceResponse:
    description: Workspace response model
    properties:
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: string
      members:
        items:
          type: string
        type: array
      name:
        description: Workspace details
        type: string
      owner_id:
        type: string
      slug:
        type: string
      updated_timestamp_utc:
        type: string
    type: object
  workspaces.WorkspacesResponse:
    description: Workspaces list response model
    properties:
      workspaces:
        items:
          $ref: '#/definitions/models.Workspace'
        type: array
    type: object
host: localhost:8002
info:
  contact: {}
paths:
...
  /workspaces:
    get:
      consumes:
      - application/json
      description: Lists all workspaces with pagination
      parameters:
      - default: 0
        description: Number of records to skip
        in: query
        name: skip
        type: integer
      - default: 10
        description: Number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspaces.WorkspacesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List workspaces
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: Creates a new workspace
      parameters:
      - description: Workspace details
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/workspaces.WorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspaces.WorkspaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - workspaces
  /workspaces/{workspace_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a workspace (owner only)
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete workspace
      tags:
      - workspaces
    get:
      consumes:
      - application/json
      description: Gets a workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspaces.WorkspaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get workspace
      tags:
      - workspaces
    put:
      consumes:
      - application/json
      description: Updates an existing workspace (owner only)
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Updated workspace details
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/workspaces.WorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspaces.WorkspaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update workspace
      tags:
      - workspaces
  /workspaces/{workspace_id}/members:
    post:
      consumes:
      - application/json
      description: Adds a new member to a workspace (owner only)
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Member ID to add
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/workspaces.AddMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add member to workspace
      tags:
      - workspaces
  /workspaces/{workspace_id}/members/{member_id}:
    delete:
      consumes:
      - application/json
      description: Removes a member from a workspace (owner only)
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove member from workspace
      tags:
      - workspaces
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"